Move BaselinesController to its own file.
